Fix for #707

+++ lib/thor/actions/inject_into_file.rb
@@ -106,7 +106,6 @@ class Thor
       # Adds the content to the file.
       #
       def replace!(regexp, string, force)
-        return if pretend?
         content = File.read(destination)
         before, after = content.split(regexp, 2)
         snippet = (behavior == :after ? after : before).to_s
@@ -114,7 +113,7 @@ class Thor
         if force || !snippet.include?(replacement)
           success = content.gsub!(regexp, string)
 
-          File.open(destination, "wb") { |file| file.write(content) }
+          File.open(destination, "wb") { |file| file.write(content) } unless pretend?
           success
         end
       end
